stack and queue fills, in differnt colors
I also fixed the algorithm. It was fine before... but not strictly
speaking a flood fill. It filled everything that wasn't the target
color, which is only suitable for fills in monochrome images. Now
it fills everything that *is* the source color (unless they are the
same in which case it does nothing).