help text + unsuccessful help display idea
Showing help this way doesn't work because there isn't enough room
below the canvas on most monitors.

In the "real" application (assuming I make something else from this
prototype), the help can be shown in a spearate window (and better
to use a WebView2 in it, which I haven't managed to get working).

But for this the best approach is probably just to have a link to
the help instead of a Show Help / Hide Help button.

Another possibility is to show the help in place of the canvas.
That may annoy users, though, in that they will have to switch
back and forth to try things out that the learn from the help.