current speed in status bar
I've also renamed "Random each time" to "Random always". This was
motivated by a desire to have more extra room in the status bar,
so it wouldn't look cramped. But that is not a sufficient reason
for the rename. I think "Random always" is less confusing. Even if
I undo the status change, I plan to keep the new name (or replace
it with something better, but not "Random each time").