set to .NET 5; fixed comment filename + no longer saying "prototype"
This doesn't work prior to .NET 5, so I'm setting the "project" to
require that, to avoid confusion.

Now the major confusion, I think, will be when people try to use it
with horizontally split panels. I don't know of any way to
programmatically detect that, so as to warn the user. (It should
be used with a vertically split view unless the screen is in
landscape orientation or the code panel is being hidden.)

This still should be a prototype... for a program with a way nicer
interface that is also portable, or otherwise has major design
advantags over this. But in the origina sense that I mean -- where
I was going to make something else in LINQPad based on this --
this is no longer a prototype.

The last "prototype" version (in the prototype branch) is still
useful for reference, since it uses the old (less OO) code
organization.