fixed ImmediateFill
Made it iterative so it does not overflow the stack.

(Note that the non-immediate recursive fill is safe from this
because it is not truly recursive--in operation--all the way, due
to the way the compiler converts it to continuation-passing style
behind the scenes. With a sufficiently high "speed" parameter it
would overflow the stack, but that would be two or more orders of
magnitude bigger than the speeds used in this program.)

Also used LockBits and UnlockBits.