trying out Charter ITC TT
I'm not sure what the license for this font is. Unlike Google
Fonts, CDN Fonts doesn't say.

It may be better to use the version available for download at

    https://practicaltypography.com/charter.html

since that is explicitly stated to be covered under a free (as in
freedom) license, even though that would require bundling it with
this program.

Also, I'm not sure I really like Charter for this program's help
file (though I'll definitely be interested in using it in some
other projects).