working WebView2 scratchwork + style fix
The stable version of WebView2 works now.

Also, the main mistake I was making was unrealed. WebView2 has to
be installed (unless one has a preview version of Edge).

I got the WebView2 runtime from:

    https://developer.microsoft.com/en-us/microsoft-edge/webview2/#webview-title

Specifically, the "Evergreen Bootstrapper".

Chromium-based browsers (including Chrome and Edge) apparently
enable ligatures by default. Normally this would be fine, but the
"fi" ligature in DM Sans (which appears in my <dt> elements) is
both unnecessary and distractingly ugly (at least when used for
display purposes as I am using it). I've turned that off.

It looks like there's no longer a good reason for me to use so many
different fonts. Switching from Crimson Pro to Source Code Pro for
the "body text" appears to have made it so that Ubuntu (the sans-
serif font) looks good for my <h4> heading. Or maybe it's some other
change--such as to the text size--that did it. It's also a bit
thicker than DM Sans, which is nice, since my <dt> elements are
conceptually headings (and are indexed as such in the ToC).

So I might change <dt> and <h4> to use Ubuntu (as <h1> does) and
look into using it for <h2> and <h3>. I feel that the sheer number
of fonts I'm using is too much, and that it makes the document feel
busy and incongruous.

But I haven't made any font changes in this commit.