fixed droplet hover regression + made hover more distinctive
Putting the droplet past the end of the <a> element to prevent it
from being underlined on hover was a bad solution, because users
will not know immediately what the droplet means and will likely
hover over it to find out. Also, it's ugly for hovering over the
droplet not to produce the hyperlink hover effect.

So I've put the droplet back in the <a> element and prevented it
from being underlined using other techniques. Now hovering over
the drop applies the underline effect to the link (but not the
droplet itself).

I've also changed background (from its near-white shade) to azure
on the whole <a> element, including the droplet. I think this makes
links more distinctive and identifiable. By itself, this would be
insufficient contrast, but it's in addition to the other effects.
This also makes it more intuitively clear that the droplet is
conceptually part of the hyperlink, an adornment on it rather than
something separate.