use Segoe UI instead of default sans-serif (Arial in WebBrowser) for tips
IE, and thus WebBrowser, uses Arial as its default sans-serif.
Edge does the same. On all recent Windows systems, the default UI
font is Segoe UI. Since that's the font that will nearly always
be showing in the app hosting the tips, using it for the tips as
well achieves greater consistency and makes the tips feel like
they are really part of the application.

It also seem to look nicer with Segoe UI, except the spacing
around it, which should be possible to fix.

The X-UA-Compatible setting to make WebBrowser behave as newer IE
so it will download the backup font (Nimbus Sans L) from CDN Fonts
creates excess space.

Either the spacing can be changed (I think this would be mostly in
the app) or X-UA-Compatible can be removed, considering that
Segoe UI is installed on Windows systems and thus the backup is
only for viewing the file on other systems (where one would not
be using IE, embedded or otherwise).