remove X-UA-Compatible from tips to fix spacing regression
It mostly fixes it. The spacing is a tad bit off still, presumably
from the new font (Segoe UI) itself.

This change may be temporary. But I want to see what it looks like
in a small (and automatically sized) WebView2 control before
deciding what the best approach is.

That could wait. But this branch should not be merged to master
until it has been tested with a display scaling of at least 200%,
preferably on a 4k screen. (I tested the old font, styling, and
geometry for the tips this way.)