abort previous runs' fills (and free memory)
This should cause the query to use fewer resources, but the real
motivation for this change is that, once (some) fills trigger
side effects that can be observed from outside the MainPanel UI
(as in the Chart branch, and as will happen with video recording),
these side effects will still occur from previous runs.

I've observed that in the chart brach; merging this should fix it.