I cringe at this failed attempt to optimize strategy comparisons.
This bad, embarrassing, awful idea doesn't work, can't work, and
due to its complexity wouldn't be worthwhile even if it did work.

I'm committing (and reverting!) this as a warning to myself and
(ugh) possibly others not to make this mistake again.

I'll also (after the revert) add a comment warning about this.

My fragile ego is bruised.

Note that the added XML comments are also not quite correct. This
code should not be reused without... well, at all.