simpler (but worse) way to trigger status updates
Because fewer messages are processed when no fill is running, this
is much less responsive than the old way in that situation.

I could combine the two approaches, updating the status both on
Application.Idle and under the circumstances I was already doing
so. I'm worried about the overhead of that approach, though

The overhead would not be much more than the overhead of doig this,
but I think this overhead may be a bit too much. It seems slightly
choppier. It's hard to tell.

If I do end up returning to this, I'll want to explore ways to
decrease the volume of memory allocations in UpdateStatus(), since
it is called so often.