instrumented recursive fill subtasks: ~75% synchronous completion
Slightly higher (~80%) for random-walk neighbor enumeration orders.

Although it's not hard to work out geometrically why this should be
the case, this information is interesting and it would be nice to
keep this instrumentation in some form. But in its current form it
diminishes code clarity. I mainly did it to verify that having the
helper return a ValueTask<> instead of a Task<> may make sense.

I may bring this instrumentation back in some form, at some point.