manual cancellation; related refactors; patched another leak
This is a fairly big change since it touches lots of the code
and it completely reworks the way cancellation works. The old way
could've been extended to support both soft and hard cancellation
(pending UI changes are performed in soft cancellation but not
hard; soft is triggered by the Stop button and hard is triggered
by disposal) by adding a third enumeration constant and handling
the different combinations appropriately, but I think this way is
even more natural, and it is also less verbose.

That will be especially relevant when charting is added, since
that's an operation with major observable effects (not just a
resource leak) that one probably still wants with soft
cancellation but definitely not with hard cancellation.

Previously, only hard cancellation was supported (and it did
work, just the code for it was big and somewhat complicated).

I'm aware of one important regression due to this commit.

At 125% desktop scaling on a 1080p screen, this pushes part of the
"Open Help" button off the screen. But I'm pushing this to master
now in spite of not yet having fixed that bug, because I think the
deduplication of starting and finishing code for the fills makes
the implementations much more understandable.

(Since much of the point of this program is to demonstrate how such
code works -- both from a flood fill perspective and from an
asynchronous programming perspective -- I think this tradeoff is
acceptable. I plan to fix this bug soon though, hopefully later
today.)