made the stop button an experimental feature (use launcher)
There are three problems with the stop button:

(1) The horizontal-space bug in the infobar, mentioned in prevous
    commit messages: part of the "Open Help" button is cut off on
    the right, on 1080p screens with 125% desktop scaling.

(2) The UI is otherwise... not very good. Mainly this is because no
    tooltip shows when the button is disabled. But there should
    perhaps be alerts some or all of the time when it is used.
    I think this can be solved without too much time and effort.

(3) The program seems less conducive to experimentation and less
    fun to use now. Cancelling fills is not nearly as edifying or
    interesting as making new ones. I'm not sure how big of an
    issue this is, but the cancel button lights up when fills are
    running and beckons to the user like the social media
    notification anti-pattern in vaguely resembles.

Perhaps showing an alert on every cancellation would clarify that
cancellation is not a feature that should be used all the time,
thereby solving the less important part of (1) and a significant
portion of (2). I'll have to try it and see.

To be clear, the alert would not tell the user to cancel less
often. It would just tell the user how many fills had been
cancelled. But since alerts are for special circumstances (and
they are shown infrequently), I think this would make my intent
clear quite nicely. And if other users want to use cancellation
more often than I think they should... then that's totally fine.

The stop button is still fully available (and fills even interact
with it, even when it's not shown) but to show it, the checkbox
for it must be seleted in the "developer mode" launcher.