fixed bug where modifier keys didn't update status when _tips was focused
This was a significant usability bug, especially considering that
the tips can sometimes be focused (if open) by clicking an area of
the PluginForm that is physically outside the tips, and because a
help feature shouldn't cause key informaton to fail to be displayed
to the user!

Nonetheless, I had not noticed this bug until today. Looking back
and testing preivous commits, it *appears* this bug has always been
present (as far back as the features it pertains to existed).