extracted Finish() parts to helpers; make tooltip appear near instantly
I'm using Uncapsulate for the tooltip, which is not great.

I've written it so it will probably degrade gracefully, and just
not succeed at changing the tooltip delay, if a change to the
library breaks it. ... Probably.

It would be good to find a better way to achieve this. The benefit
it considerable, since, without it, it's unpleasant and even
somewhat impractical to inspect more than a few specific values in
a chart of a fill that has thousands of steps.