made status text shorter and added tooltip
This is necessary to ensure the infobar fits at 125% scaling on
1080p screens, when all the buttons are shown.

Ever since adding the charting button and reclassifying charting as
non-experimental, this (or something like it) is necessary, since
125% scaling is a common default in Windows 10 on many 1080p
laptops.

The tooltip is to make up for having less explicit information in
the status text, though even the old status text could've used a
tooltip like this.

Really, a different control should be used for the status bar,
with each of the tree parts shown separately in fixed positions.