prevent selecting alert from "following the link"
This fixes the bug mentioned in point (1) in the 6507b0a commit
message.

Selecting text, then unselecting it while the mouse pointer is
still down, and then releasing the button, does have the effect of
a click. But I think that's less bad than having a situation where
users who accidentally drag the mouse pointer slightly to the left
or right and fail to click the link. (In that situation, I expect
some users to assume it's not clickable and never try again.)