non-working attempt to open panels in background and switch to them later
In light of these problems, I suspect I let a bug through into the
master branch as well, when I made the help button switch to an
open help panel. What happens when there are chart panels with
indices less than, or greater than, the help panel? I may need to
revert that for a while.