too complicated, fatally flawed alternative
The problem with this is that if you switch to another panel while
lots of charted fills are completing, you'll be switched back.

Also, it visibly switches to the charts before switching back,
which is very distracting and makes it feel like the program is a
pile of rubble that only barely manages to work at all.