warn if magnifier configuration is not found
This doesn't warn when the UseBitmapSmoothing value is absent,
which is quite normal (it's absent if the user has never configured
the Magnifier, and I believe even if they've configured it but not
checked/uncheced the box for smoothing).

But if the registry *key* is not found at all, that would
definitely bear investigation.

I don't expect that to happen; one reason I'm adding this warning
is to make the code more self-documenting. Before, it looked like I
was saying that it was normal to not have any such configuration
and that this meeant the Magnifier should be expected to work but
without smoothing.