use a stopwatch emoji instead of a bar chart emoji
The bar chart emoji was hard to understand since the bars were so
far apart (compare to the bars in the animated charting icon...and
in the actual charts), as well as just not looking very good with
this font (Segoe UI Semibold). There are two line chart emojis,
both originally representing charts of stock prices, and in this
context (at this size, in monochrome) they are inscrutible.

These are delay charts so the stopwatch seems apt. Also I think it
look good in/with this font.