use fewer fonts
This uses Ubuntu, in various weights, letterspacings, and upper
margins, for all the headings and heading-like elements.

Two problems with this make it unsuitable for merging to master at
this time.

Major Problem: The weights display correctly in WebView2/Edge but
not in WebBrowser/IE.

Minor Problem: The CSS is very disorganized, with commented-out
code, and likewise the stylesheet code in the HTML document
probably pulls in some styles that are not being used.

The minor problem is not hard to fix. The major problem makes it
seem unlikely I will ever actually use these changes. I'm committing
them on a feature branch anyway so I don't lose them, in case
there's a way to make this work properly.

Also, some ideas here should probably be used, like the bigger
margins above headings.