added XML comments to the types I plan to use
NB I don't plan to use these types in exactly this form. The code
to deal with resizing takes up a lot of space and I don't want to
have to test (and then properly maintain) it. So I'll probably
just throw an exception if either the the host or the child are
resized. That's unsuitable for general use, since event handlers
could still be called when the child is resized, before the
exception is thrown. But for my purposes (I know specifically how
I'm using it) I think that should be fine.