pinned WebView2 to 1.0.664.37 as 1.0.705.50 drops EdgeNotFoundException
This is a temporary measure, probably. I need to find out why
EdgeNotFoundException is no longer supplied, and what happens under
that condition with the newer version.

Ultimately, all nuget packages will be pinned (to protect against
this sort of thing) and then tested periodically against newer
versions and updated where approprite (which will be usually). But
since there is no stable version of Flood yet, I haven't bothered
to do that, preferring to face and deal with breakages like this
one as they come.