made help look better on mobile
Since the help is mainly intended to be browsed on a desktop/laptop
computer on which one is running the software -- and the software
doesn't work on mobile devices -- this is not very important. But
it's no harm.

Before this, some mobile devices would display the text at full
size, which was okay but a bit bigger that necessary given the
distance at which readers tend to hold their mobile devices (and
given the ability to zoom easily while still seeing the full width,
when margins are visble). But other mobile devices would, at least
going by the results in Firefox's Responsive Design Mode, show the
text very small.

Now the text should consistenetlyappear moderately smaller than on
a regular browser, but still big enough.