tips navigation to help side-highlights most relevant info
By "most relevant info" I mean whatever help contents is most
directly related to the tips item that has been clicked.

Sometimes this is the same as the whole (sub)section whose heading
has the anchor. Other times, it is some smaller part of that
(sub)section.

Before, tips links went to anchors on the most relevant info. Now,
they always go to the most specific (sub)section that contains it.
When those are different, the highlighted portion begins inside
the (sub)section, lower than the very top. I think this is clearest
and most intuitive behavior, in practice.