fixed unintentional :after vs. ::after inconsistency in IE hack
The hack does not depend on, and is not affected by, this. It
worked before and still does (and it is also still necessary).

It was confusing to have them different (one with the pseudoelement
notation and the other with the pseudoselector notation), since the
hack works around one very strange behavior by relying on another
very strange behavior, and one might understandably have assumed
that a third very strange behavior was somehow involved.