removed icomoon stylesheet <link> and optional navlink-glyph TODO
I never meant to commit a version that pulls in an icomoon
stylesheet. No commit (or even a non-committed version of the
file) has actually used anything from icomoon, though I was
considering it.

I don't expect to use a chain-link image/glyph for the h1 navlink,
because I think the current circular arrow is adequate, and a
picture that signifies a hyperlink in the user's mind is really not
all that useful here. All the other navlinks have a glyph that is
meaningful in a *specific* way, so this one should too (and does).