scroll to current section when going there again from tips
Without this, clicking a link in tips, scrolling manually in the
help window (as users will often do), and then clicking that same
link in tips again, returns to the section anchor in IE/WebBrowser
but not in Edge/WebView2.

I might end up going with another approach though.

See comments for details.