fixed first-scroll bug with WebBrowser/IE
The bug this fixes is that, with a WebBrowserHelpViewer, clicking
a link in tips to open a section in the help viewer, and then
closing the help viewer panel, and then clicking it (or, I think,
another link) in tips, would cause the panel to reopen but to
scroll to the wrong position. I believe there were other sequences
of interactions that also caused WebBrowser/IE to fail to scroll
to correct position. All involved opening a new help viewer.
Reusing a help viewer that was still open was not affected.

Note that this is essentially the opposite of the other scrolling
bug that affected WebView2/Edge, where *subsequent* navigations to
the same fragment didn't work. Here, it was the initial navigation
that didn't scroll correctly.

(Note also that this is totally unrelated to the *other* bug with
WebView2/Edge where closing the panel would throw an exception.
That bug was due to a bug in WebView2, recently fixed in a pre-
release, and upgrading to that pre-release allowed my workaround to
be removed.)