bad non-working attempt to delete temp dir
This would probably be a bad design even if it did work. But it
doesn't work, because the process accessing the user data folder
is still running at the time of disposal.

I'm reverting this immediately and only committing it in case I
need something from it later (and to document why it doesn't work).
A better approach will probably be one of the following:

(a) Use System.Diagnostics.Process to wait (starting early) on the
    process asynchronously and delete the folder when the process
    exits.

(b) Have subsequent WebBrowserHelpViewer objects (or the factory
    method that makes them) take care of deleting old user data
    folders. But I don't like this much, since doing it correctly
    would be hard. I don't want to break the ability to have
    multiple help panels open. (The program doesn't do this...
    but it will happen if the user runs multiple instance of the
    program, which would probably be from multiple instances of
    LINQPad. This is something that users could do intentionally
    and I don't want to break it. Furthermore, I don't want to
    bother detecting it--even if I do break it, having weird things
    happen in this situation, without any explanation to the user,
    is not acceptable.)