second Hyphenopoly attempt - jsdelivr instead of cloudflare
This does work... in WebView2/Edge, where it is needed. But it
breaks WebBrowser/IE, which doesn't need it anyway, because IE
doesn't support the => syntax, which appears in the loader script!

Although Hyphenopoly doesn't need to work in WebBrowser/IE, this is
breaking JavaScript in it, so the current situation can't be kept.