fixed first-scroll regression (?) with WebBrowser/IE
Commit 5985e0f fixed a bug where a new WebBrowserHelpViewer would
often fail to scroll to the fragment speified in its hash, by
working around (what I strongly presume to be) the underlying IE
bug using JavaScript to call scrollIntoView() on the element. Or so
I thought...

But either that fix was incomplete, or (and I think this is what
happened) something else caused the bug to come back. I don't know
what is cauasing it, so it's hard to be sure I've really fixed it
now. But further delaying the programmatic scrolling operation
until the next cycle of the event loop seems to work.