script cruft removal + refactoring; added missing comments
This documents the IE initial-load workaround situation.

The refactor is mainly for clarity but it also makes manual testing
easier by giving individual lines for both major actions that can
be temporarily commented out. Most added comments are in the script
but I also fixed an empty <remarks> section in a C# XML comment.

Also, I had previously changed #use-the-wheel-luke to #scrolling
when it was expanded to envelop more text, since that caused it to
be other user-input aspects of scrolling (e.g., modifier keys) too.

That was a good change, which I'm keeping, but I've re-added
the #use-the-wheel-luke id to the <p> that originally had it, as an
easter egg.