removed scratchwork for tips approach I didn't take
This was for attempting to detect key-up in WebView2 so it could be
used as an alternative tips browser. My main goal wasn't to really
use WebView2 for this purpose (unlike the help browser, WebBrowser
is quite satisfactory for the tips browser) but to make it so that
HelpViewer would work for both the tips and help browsers.

It started to look like a lot of work, involving COM, and also it
was unclear that this design approach even made sense, so at least
for now, I am not attempting to unify the two.