fixed some minor problems static analysis found
(1) Repeated words in XML comments.
(2) Unnecessarily interpolated strings.
(3) Instance members that could be static.
(4) An unused method that could be removed.

Analysis was on the csproj-scratchwork branch in VS Code with
OmniSharp, running Roslynator analyzers.

It complained about way more, but these are the things that I know
were actual problems. (I haven't yet configured which analyers I
want for this project.)