a JavaScript loading overlay is unlikely to work
I'm not totally sure, but it looks like the WebView2 lag, though
some of it is brought about by the JavaScript that I've added
(perhaps especially Hyphenopoly), is happening early enough that
a loading overlay implemented in JavaScript would be hard to make
appear early enough.

I also don't want to slow down the page further to get it to show
a loading overlay.

The loading overlay in this commit does not really work, because
there is no code to dismiss it when the page is loaded. It's just
a demonstration of why I don't think a JavaScript-based loading
overlay is my best option here.

This commit is for reference. I'll be reverting it immediately.