improved tips fallback readability
Windows systems will have Segoe UI, but this is to improve
readability when tips.html is loaded in an ordinary web browser
(which could be on any OS).

I want this to look okay mainly because tips.html is a handy
higlight-applying portal to help.html, even outside the application.