put Nimbus Sans L back as tips font fallback
It turns out the reason it didn't look good was that the CDN wasn't
providing bold properly. I don't know how to fix that--looking at the
font-face declarations makes it seems like that should be working.
Anyway, it works well as a fallback on systems that already have it,
so I put it back, removed the link element that tried to pull it in
from a CDN (note: eventually, no fonts will be gotten by CDN anyway),
and kept DM Sans as a second fallback since it looks better than the
others.