fixed tips kbd sizing to be consistent across browsers
Before, kbd elements in tips.html had displayed as desired in
WebBrowser/IE, but they were too big in other browsers.

This fix is tested (in the application) with 100% and 125% display
scaling. Setting it to 10px didn't scale up properly (as expected)
but setting it to .75em works well across browsers, and it does scale
properly (at least in IE/WebBrowser in the application).