set tips size via query string in app
This moves the "small" font size from CSS into the URL the
application uses for the tips. This way, the tips appear at a more
reasonable size when viewed in a web browser (where they are useful
since they have nice section and "topic" links).

I tested this with 100% and 125% scaling and it seems to work fine.

The code change also leads the way for a possible future feature
where the tips size is configurable. (I don't know if I'll implement
that though.)