added Nimbus Sans after Nimbus Sans L
It looks like, on Ubuntu, Nimbus Sans L is not... installed? Firefox
uses it. Chrome doesn't. Nimbus Sans doesn't look as good as Nimbus
Sans L for the tips, because the letters are shifted upward. I don't
know of a way to fix that that wouldn't mess things up for other
fonts (and I don't think it would be worth increasing complexity at
all in the attempt).

But Nimbus Sans still looks okay enough (and much better here than
Liberation Sans, as well as better than actual Arial) that I think
it's worth putting it in the stack, and before the fonts that look
way different.