added missing 2020 to license header on files that started then
Also added it to LICENSE itself.

The modified files, and LICENSE, now give "2020, 2021" as the
copyright years, rather than just "2021". This omission was a
mistake. Note that I haven't made this change in all files,
because some files' entire contents were written in 2021.

This project, and some but not all of the files in it, started out
in late 2020. (Some files that appear, with some tools, to have
been created in 2021, really contain material from 2020, as
revealed by running "git blame" manually and inspecting/searching
the output.)