remaining dependency licenses in help + fixed some mistakes in readme
This puts all the dependency information in the help that is
present in the readme.

I've removed the TODO that said to get rid of the Source Code Pro
dependency, since I'm using that font for license texts in the
help. It looked better than anything else I tried for that, perhaps
due to pairing well with Source Serif [Pro] (which is used for
normal/body text in the help).

There's a major bug in the new help material: IE/WebBrowser doesn't
support the details element, so all licenses are shown in full (and
uncollapsable) when viewed in the application if WebView2 is
unavailable (or otherwise unused).

I think support for the details element can be polyfilled.